% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-speedtest-cli.R
\name{install_speedtest_cli}
\alias{install_speedtest_cli}
\title{Provides instructions for installing the official speedtest CLI application}
\usage{
install_speedtest_cli()
}
\description{
Provides instructions for installing the official speedtest CLI application
}
