% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{nice_speed}
\alias{nice_speed}
\title{Convert a test speed, in Mbits/s, to its string representation along with
appropriate units for the magnitude of the test speed}
\usage{
nice_speed(number)
}
\arguments{
\item{number}{numeric The speed to be converted}
}
\value{
character The character representation of the speed
}
\description{
- Assumes number is in Mbits/s to start off with
- Only chooses between Kbit/s, Mbit/s, and Gbit/s
}
