% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestest.r
\name{spd_best_servers}
\alias{spd_best_servers}
\title{Find "best" servers (latency-wise) from master server list}
\usage{
spd_best_servers(servers = NULL, config = NULL)
}
\arguments{
\item{servers}{if not \code{NULL}, then the data frame from \code{\link[=spd_servers]{spd_servers()}}. If
\code{NULL}, then \code{\link[=spd_servers]{spd_servers()}} will be called to retrieve the server list.}

\item{config}{client configuration retrieved via \code{\link[=spd_config]{spd_config()}}. If \code{NULL} it
will be retrieved}
}
\value{
server list in order of latency closeness (retrieval speed column included)
}
\description{
Find "best" servers (latency-wise) from master server list
}
\note{
the list of target servers will be truncated to the first 10
}
