% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest.r
\name{spd_closest_servers}
\alias{spd_closest_servers}
\title{#' Find "closest" servers (geography-wise) from master server list}
\usage{
spd_closest_servers(servers = NULL, config = NULL)
}
\arguments{
\item{servers}{if not \code{NULL}, then the data frame from \code{\link[=spd_servers]{spd_servers()}}. If
\code{NULL}, then \code{\link[=spd_servers]{spd_servers()}} will be called to retrieve the server list.}

\item{config}{client configuration retrieved via \code{\link[=spd_config]{spd_config()}}. If \code{NULL} it
will be retrieved}
}
\value{
server list in order of geographic closeness
}
\description{
Uses \href{https://ipinfo.io}{ipinfo.io} to geolocate your external IP address.
}
