% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth.r
\name{spd_compute_bandwidth}
\alias{spd_compute_bandwidth}
\title{Compute bandwidth from bytes transferred and time taken}
\usage{
spd_compute_bandwidth(size_bytes, xfer_secs, mbits = TRUE)
}
\arguments{
\item{size_bytes}{size (in bytes) of the payload transferred}

\item{xfer_secs}{time taken for the transfer}

\item{mbits}{produce output in megabits (Mb)? Default: \code{TRUE}}
}
\description{
Compute bandwidth from bytes transferred and time taken
}
\examples{
spd_compute_bandwidth(19200000, 1) # 150 Mb/sec
}
