% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.r
\name{spd_config}
\alias{spd_config}
\title{Retrieve client configuration information for the speedtest}
\usage{
spd_config()
}
\description{
Retrieve client configuration information for the speedtest
}
\examples{
\dontrun{
spd_config()
}
}
