% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers.r
\name{spd_servers}
\alias{spd_servers}
\title{Retrieve a list of SpeedTest servers}
\usage{
spd_servers(config = NULL)
}
\arguments{
\item{config}{client configuration retrieved via [spd_config()]. If `NULL` it
will be retrieved}
}
\value{
data frame
}
\description{
Retrieve a list of SpeedTest servers
}
