% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-main.r
\name{spd_test}
\alias{spd_test}
\title{Test your internet speed/bandwidth}
\usage{
spd_test()
}
\description{
No-frills. Picks the closest geographic server with least latency,
performs download & upload tests and returs the best result.
}
\details{
Make a command-line alias or executable with:

\verb{Rscript --quiet -e 'speedtest::spd_test()'}
}
\examples{
\dontrun{
spd_test()
}
}
