% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedtest-cli.R
\name{speedtest_cli}
\alias{speedtest_cli}
\alias{print.speedtest_cli_result}
\title{Perform an official Ookla Speed Test via their command line tools}
\usage{
speedtest_cli(progress = TRUE)

\method{print}{speedtest_cli_result}(x, browse = FALSE, ...)
}
\arguments{
\item{x}{`speedtest_cli_result` object}

\item{browse}{if `TRUE`, open a browser window with the results}

\item{...}{ignored}
}
\description{
Perform an official Ookla Speed Test via their command line tools
}
