% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers.r
\name{spd_servers}
\alias{spd_servers}
\title{Retrieve a list of SpeedTest servers}
\usage{
spd_servers(config = NULL)
}
\arguments{
\item{config}{client configuration retrieved via \code{\link[=spd_config]{spd_config()}}. If \code{NULL} it
will be retrieved}
}
\value{
data frame
}
\description{
Retrieve a list of SpeedTest servers
}
\examples{
\dontrun{
config <- spd_config()
spd_servers(config)
}
}
